Restructured, separate classes/files for each entity
